Now registers the service within GraphQL to skip the attribute. :) Time to get in Postgres.

+++ bookmanagementapi/bookmanagementapi/Application/Abstractions/IBookRepository.cs
@@ -5,7 +5,7 @@ namespace BookManagement.Application.Abstractions;
 public interface IBookRepository
 {
     IEnumerable<Book> GetAll();
-    Book? Get(int id);
+    Book? Get(long id);
     IEnumerable<Book> GetByAuthorName(string authorName);
     IEnumerable<Book> GetByTitle(string title);
 }
